liten refakturering: flyttet beloep formatering til utils
