+++ app/src/pages/skjema/skatteforhold-og-inntekt/SkatteforholdOgInntektSteg.tsx
@@ -27,6 +27,10 @@ import {
   Skjemadel,
   type UtsendtArbeidstakerSkjemaDto,
 } from "~/types/melosysSkjemaTypes.ts";
+import {
+  formaterBelopForVisning,
+  stripBelopFormatering,
+} from "~/utils/belopFormat.ts";
 import { getFieldError } from "~/utils/formErrors.ts";
 import { useTranslateError } from "~/utils/translation.ts";
 
@@ -46,42 +50,6 @@ type SkatteforholdOgInntektFormData = z.infer<
 interface CheckboxAlternativ {
   verdi: string;
   label: string;
-}
-
-/**
- * Formaterer en rå inputverdi som et norsk kronebeløp for visning.
- * - Erstatter punktum med komma
- * - Avrunder til 2 desimaler
- * - Legger til tusenskilletegn (mellomrom)
- * Eksempler: "1234.456" → "1 234,46", "1000000" → "1 000 000,00"
- */
-function formaterBelopForVisning(value: string): string {
-  const trimmed = value.trim();
-  if (!trimmed) return "";
-
-  // Normaliser: erstatt komma med punktum for parsing
-  const normalized = trimmed.replaceAll(/\s/g, "").replace(",", ".");
-  const parsed = Number.parseFloat(normalized);
-  if (Number.isNaN(parsed) || parsed < 0) return value;
-
-  // Avrund til 2 desimaler
-  const rounded = parsed.toFixed(2);
-  const parts = rounded.split(".");
-  const helTall = parts[0] ?? "0";
-  const desimaler = parts[1] ?? "00";
-
-  // Legg til tusenskilletegn
-  const medTusenSkille = helTall.replaceAll(/\B(?=(\d{3})+(?!\d))/g, " ");
-
-  return `${medTusenSkille},${desimaler}`;
-}
-
-/**
- * Fjerner formatering (mellomrom) fra et visningsformatert beløp,
- * beholder komma og siffer — formatet melosys-skjema-api forventer: "1234,46"
- */
-function stripBelopFormatering(value: string): string {
-  return value.replaceAll(/\s/g, "");
 }
 
 function SkatteforholdOgInntektStegContent({
